+++ docs/source/whats_new.rst
@@ -21,6 +21,7 @@ Enhancements
 ~~~~~~~~~~~~
 
 
+- Add GradMax normalization details and associated tests for update scaling behavior (:gh:`242` by `Pako Maxence TEKOU`_)
 - Introduce VGG-like growing architecture support with the `VGG` class (:gh:`236` by `Hugo Mousset`_)
 - Adds support for selecting GroupNorm as the normalization layer in the ResNet container, alongside the existing BatchNorm option (:gh:`233` by `Théo Rudkiewicz`_).
 - Adds configurability to the ResNet container to allow using BatchNorm2d or disabling normalization entirely (:gh:`228` by `Théo Rudkiewicz`_)
@@ -134,3 +135,4 @@ API changes
 .. _Stéphane Rivaud: https://github.com/streethagore
 .. _Félix Houdouin: https://github.com/Edarfix
 .. _Hugo Mousset: https://github.com/hmousset
+.. _Pako Maxence TEKOU: https://github.com/maxencelebaron